3. Reflection Agent/requirements.txt
openai
gradio
python-dotenv
pyyaml
